Created TicketServiceImpl Logic
Added Booking, BookingValidator, TicketpriceCalculator

+++ cinema-tickets-java/src/main/java/uk/gov/dwp/uc/pairtest/utils/TicketPriceCalculator.java
@@ -4,5 +4,5 @@ import uk.gov.dwp.uc.pairtest.domain.TicketTypeRequest;
 
 public interface TicketPriceCalculator {
 
-	float calculate(TicketTypeRequest request);
+	int calculate(TicketTypeRequest request);
 }
